+++ packages/observe/unix/write.mli
@@ -1,0 +1,3 @@
+type writer = Unix.file_descr -> string -> int -> int -> int
+
+val all : write:writer -> Unix.file_descr -> string -> unit